feat: auto-updater for vscode extension (#68)
Upon extension activation (opening a .w file for the first time in a window) the extension will check the latest release with a certain tag (`development` by default) from the winglang repo. It gets the checksum from the latest vsix file there. If that checksum is different than the one currently stored in vscode's persisted state it will download the new vsix, install it, and ask the user to reload the window. 
This checking process is rate limited, currently set to have a minimum of 1 hour between checks.

A github PAT is needed for this functionality, so I added a setting for that and the option to customize the tag cause why not. In the future we should have a proper auth flow setup, but I took the shortcut for now.

TODO for future PRs:
- This project really needs projen
- If you never open a new vscode window it will never check again. Need to put the update on a polling schedule.
- Need to have a command to update manually
- Proper Github auth
- We're using a node API to download files, would be nice not to be node-dependent