feat: add language server project and workflow (#78)
This adds a rust project for a language server. I used
[tower-lsp](https://github.com/ebkalderon/tower-lsp) as a nice framework
to handle the boilerplate of a language server, so the shape of the code
here is mostly due to that. I went ahead and added some extremely
incomplete features:
- semantic token highlighting
- Detecting when an identifier is a variable vs other contexts.
Incomplete.
- completions
  - some keywords, able to be shown kinda contextually
  - variables, but doesn't handle scoping properly
- diagnostics
  - tree-sitter errors only
- hover
- Shows current AST node and ancestors. This is purely a debug thing to
be removed later.
<img width="395" alt="image"
src="https://user-images.githubusercontent.com/1237390/189220007-4932eb2c-ba81-4f5b-a44e-390d04a49804.png">


This also sets up the workflow to build and include the server in the
vscode plugin, which registers a client using that binary. Lots of rough
edges but I think it'll work.

Co-authored-by: Chris Rybicki <chrisr@monada.co>