feat: combine inflight code and simulator code into wingsdk (#103)
The goal of this PR is to combine all code related to implementing Wing SDK resources into a single repo to make it easier for external contributors to contribute and change resources in the future.

Previously I couldn't see an easy way to accomplish this since if we naively included inflight client classes with preflight classes, then:
1) jsii would impose additional restrictions on TypeScript code that did not seem to be necessary since the typescript inflight clients are not meant to be transliterated to other languages
2) jsii limits the dependencies that can be used in TypeScript projects since it uses an older version of TypeScript

@eladb had some insights that let us address both of these. (2) is addressed by bundling all non-jsii dependencies (for now, I think using `bundledDependencies` should be sufficient?). (1) is addressed by switching compilation into two passes - one pass that uses jsii to compile preflight code, and one pass that uses a new version of TypeScript to compile inflight and simulator implementation code.

The work here mostly aligns with the design from https://github.com/monadahq/rfcs/pull/47. 

I've also updated the naming in several places to switch from "local" to "sim" or "simulator".

Supersedes #95
Fixes #91

TODO (may be in subsequent PRs):
- Add unit tests for inflight clients
- Add unit tests for simulator implementations
- Rework local simulator implementations (not sure if trpc is still needed?)
- ~Set up eslint rules so that there are no imports between preflight modules, inflight modules, and simulator modules~ (done)