feat: simulator implementations for Bucket and Function (#106)
Rework the simulator implementations around the concept of "init" functions. This structure lets us easily unit test APIs in TypeScript/JavaScript without the need to spin up a REST API server. (In the future it's possible we can create a REST API that calls these classes under the hood - but it's not a MVP requirement right now)

Closes #88