chore: refactor app component into layouts (#48)
Now, the App component will be responsible of instantiating the TRPC client, the TRPC and query providers, and also listen to any schema changes. It will also render the VSCode layout (the light themed one).

I'd like to offer a switch to the Unity layout (the newer, dark-themed one) as well soon. This refactor will allow that.