fix: Implement Send for AST structs with raw pointers for use in async (#147)
In order to use these structs in an async context, TypeRef and TypeEnv
needed to implement send. I'm not actually sure if this is bad, or even
if my comments are unwarranted and this is actually totally ok. Let me
know if you know.

On the bright side, it does allow me to use `Scope` in the language
server :)