fix: add diagnostics for recoverable type checker errors (#264)
This follows a similar pattern for parser diagnostics. The type checker
will do its best to type check the whole program instead of panicking
when you try to do something like add a number and a struct.

Plenty more to do here on multiple fronts, but this at least prevents
panics for the "mostly-correct" programs and makes it easier to use the
type checker in the language server.


Side note: I'm now aware of https://github.com/Manishearth/elsa thanks
to @Chriscbr and I think this would be a better way to manage diagnostic
data. A refactor to use it would be good at some point

Co-authored-by: Chris Rybicki <chrisr@monada.co>