fix(wingii): optimized performance of wingii (#277)
This PR addresses the performance bottleneck of wingii. After evaluation
of the flamegraphs below, I spoke with @Chriscbr and it turned out for
our compiler's purposes, recursively loading a JSII manifest is not a
requirement. Since wingii is a port of the original jsii-reflect npm
package, it carried that functionality with it. With this change,
compiler can now opt out of that behavior.

On my system, the following improvement is observed (reduction of about
86% on wingsdk's JSII manifest):

```sh
# old behavior
$ time cargo run -- ../../examples/simple/sdk_capture_test.w
cargo run -- ../../examples/simple/sdk_capture_test.w  7.60s user 0.43s system 100% cpu 8.014 total

# opt out behavior
$ time cargo run -- ../../examples/simple/sdk_capture_test.w
cargo run -- ../../examples/simple/sdk_capture_test.w  1.00s user 0.19s system 104% cpu 1.137 total
```

Flamegraph of old behavior:


![image](https://user-images.githubusercontent.com/5657848/196055588-ee58e93a-d2f4-4bca-bf79-46b6bba0c5e2.png)

Flamegraph of new behavior:


![image](https://user-images.githubusercontent.com/5657848/196055661-21a61042-7cd6-4503-8001-fc56afc76b5a.png)

The flamegraphs indicate that we still have the same divide between
wingc and serde code, which is good! it means we are not leaking memory
or have other hidden costs somewhere. It's simply a matter of # of serde
calls.