fix: add env to optionally force jsii imports to be `anything` (#274)
With the recent addition of "bring cloud" working, wingc assumes you have the wingsdk package available (specifically, at `../wingsdk`).

This makes wingc unusable in a browser context without some additional work (e.g. preloading the WASI filesystem or providing a way to load up NPM modules). This is doable, but not in the current sprint. We still want the playground to work though, so here's a hack.

In addition, JSII imports are currently very slow. This makes wingc typechecker unusable via the language server. So until then the language server will use this hack as well.

This hack is not foolproof. Types are needed during jsify for certain situations, having JSII ignored causes invalid output.