feat(cli): compile and run commands (#330)
This PR is in continuation of #289 and #228 and addresses #40. It
implements `compile` and `run` commands according to the RFC in #322

Some notes:

- `cdktf` execution happens in-process in `node:vm`
- I ended up not sharing strings from Rust to WASM as I did in #228 .
Now data is passed to `wingc` and `wingc` is treated as a consumer
rather than producer of strings (resulted in cleaner code).
- This PR assumes the Wing Console is reachable through the
`wing://run?path=...` deep link already registered with the OS through
console's installation. When it's not installed, currently it pops a
dialog like this:
![image](https://user-images.githubusercontent.com/5657848/197926900-38c0eddc-ea24-4052-a218-9901a10a25ca.png)
- As to why this is not implemented in Rust, if you look at the CLI RFC
in #322, if CLI parsing happens in Rust, implementation would be split
50/50 between Node and Rust (`watch` and `eject` commands are Node
dependent and `pkg` works nicer this way after bundling). This way,
`wingc` is treated as a library and CLI parsing is done in Node to
contain the logic in a single place.

Co-authored-by: Elad Ben-Israel <eladb@monada.co>