chore: add logo to readme (#388)
Add logo and some initial badges (later on we will add version, npm, etc
but for now we are limited because things are not public).

I think we might need to change the logo images so that they don't have
so much padding around them. So we can control the spacing as we place
them. copy @amielpb