Breadth first AST scanning in type checker.
* Inner scopes now aware of all symbols and types in the outer scope regardless of where outer symbols were defined.
* TypeEnv now aware of where types where defined in the scope we can decide whether to take that into account in lookups.
* Types, namespaces, methods and properties are treated as if they are defined before other statements (implicit forward declaration).

The aim here is for an expression that's build of an inner scope to know the type of itself when being evaluated (for anonymous closure recusion, or `if` expression type-checking if we decide to make `if`s expressions, etc).

This also provides some support for imlicit forward declarations (as long as they are accessed from an inner scope).