fix(cli): upgrade now works as expected on mac (#595)
`tsc` incorrectly treated `is-installed-globally` as an actual ESM
import instead of a synthetic default import. as a result, everywhere
this value got passed, instead of a `boolean`, an `undefined` got thrown
around. As a result, this lead to a subtle bug in the `update-notifier`
package to cause it function as expected on Linux but break on Mac. All
of this happened because of incorrect `d.ts` for
`is-installed-globally`.

Co-authored-by: Uri Bar <106860404+staycoolcall911@users.noreply.github.com>