feat(sdk): tracing api (#544)
Introduces a tracing API for the simulator to make it easier to inspect the behavior of an application. At any point during a simulation, a resource can call `addTrace` or `withTrace` via the simulator `context` object to add traces and logs. Traces are breadcrumbs of information about resource operations that occurred during simulation, and can include information about cloud resources or user logs.

A list of all traces, logs, or events (both) can be obtained by calling `listTraces()` on a `Simulator` instance. Every event includes a message, a timestamp, a resource path, and a "type" field indicating what type of event it is.

By calling `onTrace` on a `Simulator` instance, a subscriber can also be added that will be called each time a new trace is emitted.

Future work: adding support for request IDs.

Closes #378