chore(sdk): rename `.sim` to `.inflight`
Since our simulator is now single process in-memory, the inflight clients and the actual resource are implemented together in the same class and they are all basically inflight.

Rename all the `.sim.ts` files into `.inflight.ts` files and clean up the special-casing that was required for `.sim` files in build.

Misc: consume `polycons` instead of `@winglang/polycons` now that it is publicly available.