fix: changing .wsim files' logo (#807)
link to the issue: https://github.com/winglang/wing/issues/503

* I chose the "easiest path" to complete this task, (and as long both
the ".w" and ".wx" files share the same icon, alias and configuration,
that seems like the best approach) I could also create a separate
VSCodeLanguage object, to the executable files, but I couldn't know its
aliases and configuration (only the only optional icon property 😂) and I
don't know how better practice it is :)

* I also struggled a bit with the "releasing" of the package locally, so
I've added a few lines to the readme file, please correct if needed

* The other change is adding the .vscode/settings.json to the git
ignore, since it added it when I added my token (as written in the
readme) and that's probably shouldn't go on git :)

It was super fun, and the documentation is awesome! (form all the dozens
of time I happen to install wing, that was definitely the fastest! I got
it up and running in less than 5 minutes) the design is smooth and there
are not a lot of words on each page (so even super tired and jet lagged
developers can read it successfully)
I'll definitely be back for more small tasks ;) 

*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

Co-authored-by: tsuf98 <tsuf@equitybee.com>
Co-authored-by: Mark McCulloh <Mark.McCulloh@gmail.com>