feat!: support new simulator file format (#222)
Previously, a simulator file (.wx) was internally modeled as a JSON file that provided information about the construct tree AND resource configuration needed to simulate the app. Now, the simulator file (.wsim) only includes information needed to simulate resources.

Full information about the construct tree is now in the tree.json file, which is generated separately. This file is generated with the intent of compatibility with the tree.json file format produced by AWS CDK. 

Fixes #221.