feat(spec): use `inflight` modifier instead of `~`
Replace the `~` symbol with `inflight` to make it more approachable and intuitive.

This applies both for closures:

    let handler = inflight () => { ... };

And for resource members:

    resource Foo {
      inflight bar() { ... }
    }

This is based on feedback from the community (credits to @skinny85) and the mental model where `inflight` is a special type of `async`.

Update all documentation, grammar and tests.