feat(compiler)!: use `inflight` modifier instead of `~` (#866)
Replace the `~` symbol with `inflight` to make it more approachable and
intuitive.

This applies both for closures:

    let handler = inflight () => { ... };

And for resource members:

    resource Foo {
      inflight bar() { ... }
    }

This is based on feedback from the community (credits to @skinny85) and
the mental model where `inflight` is a special type of `async`.

Update all documentation, grammar and tests.

Resolves #878

*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

BREAKING CHANGE: Inflight functions are now defined using `inflight ()
=> {}` instead of `() ~> {}`. Inflight methods are defined as `inflight
method() { }` instead of `~ method() { }`. This is to improve
readability and approachability.