feat(sdk): prepare to support user-defined resources (#869)
This PR has internal API changes and groundwork needed for the Wing compiler to support user-defined resources (#542). By itself, this PR does not introduce any new features or capabilities.

The bulk of the changes relate to the internal `_bind` method implemented on all SDK resources (a method not currently exposed as an API in Wing). This method plays the important role of gluing resources together when a resource is used inflight.

The method has had its signature updated to now take two parameters: the first is a host resource, and the second is a list of inflight operations that are expected to be called on the resource. The `_bind` method is also no longer returns the resource's inflight code, and instead this responsibility has been split into `_toInflight()`.

The motivation of these changes are two-fold:

1. Eventually allow the `Inflight` class to be deprecated. We think it we might be able to simplify the SDK and the compiler by modeling everything as a resource, so that for example `queue.onMessage` accepts a resource with a `handle` method instead of a special data type.
2. Allow the compiler to model resources with complex permissions. This can be accomplished through the SDK-internal `_annotateInflight` method, which lets you encode information like "If this resource's `put_twice` is called, then I will need permissions to perform `put` on `this.bucket`:

```ts
MyResource._annotateInflight("magic_put", {
  "this.bucket": { ops: ["put"] },
});
```

This PR also updates several parts of the SDK to remove the use of the word "capture", in favor of "bindings" and "bound resources".

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.