feat(sdk): display `print()` in console (#1012)
The Wing compiler converts all `print()` calls to `console.log()` calls
(both in preflight and inflight), so we attach the `console.log`
function to the `Logger` instance for all inflight handlers.

To achieve this:
1. We register a `Logger` for all apps.
2. We bind the symbol `$logger` to this logger in preflight.
3. We override `console.log` to call `$logger.print`.

NOTE: The `print()` method of `ILogger` is now a synchronous method (all
inflight functions are async by default). Not sure exactly what the
implications of this are and if it's a safe assumption but for now it
seems like it will work.

Refactored the generated code for inflight handler so it will be
possible to override `console.log` by accessing `clients` outside of the
handler object. I am wondering if this is something we will need to
revisit when we have inflight methods for resources (copy
@yoav-steinberg).

Resolves #991
Resolves #591



*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>