feat(sdk): `print`s are displayed in wing console (#1023)
All applications now have a `Logger` registered to them.

In order to allow inflight code to emit logs, we *implicitly bind* the
logger into the function code and wire `console.log` to `logger.print`.
This means that now, all `console.log` logs in inflight will be sent to
the logger.

In preflight, we still just emit `console.log` instead of `print`. At a
later stage it might make sense to actually call
`Logger.of(this).print()` (in *preflight*) instead.


NOTE: The `print()` method of `ILogger` is now a synchronous method (all
inflight functions are async by default). Not sure exactly what the
implications of this are and if it's a safe assumption but for now it
seems like it will work.

Refactored the `Function` classes so that all common code (the stuff the
produces the handler asset) is now in `FunctionBase` (as well as
handling environment variables). This is where I could implement the
common behavior that captured `console.log` and mapped it into the
logger.


Resolves #991
Resolves #591

*This is a redo of #1012 which was reverted by #1020 and fixed via [this
commit](https://github.com/winglang/wing/pull/1023/commits/093d6d23a4e3d765a97342808c65b2e7111f6c97).*

*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>