feat: capture immutable values
It is now possible to capture immutable primitives (strings, booleans, numbers) and collections (array, set, map). This is done by serializing them to JSON and including in the inflight bundle.

SDK support:
In order to support this in the SDK, two things needed to happen.
First, the base `resource._bind()` method now only calls `_bind()` on the bound resource (and adds a connection) if it has such as method. Otherwise, it simply passes it through. This means that SDK users can technically pass any object and the SDK will not be defensive about it.
*We currently assume "capture-safety" is the responsibility of the Wing compiler.*
Similarly, we only call `toInflight` if the object has this method. Otherwise, we simply call `JSON.stringify`. There's a special case for `Set` because JSON.stringify doesn't work out of the box for sets.

Misc:
* Added a couple of tests: `capture_collections.w` and `capture_primitives.w`.
* Propagate a mutable `Diagnostics` vector into `capture.rs` so it is possible to report errors. Specifically, if users try capture an uncapturable type, an error is emitted.
* The SDK now takes a snapshot of the inflight code. To avoid ugly paths in the code bundles, I am filtering out any lines that start with `//`. Not ideal, but harmless (famous last words).

Fixes #1061