chore: alerting workflow triggered by check_run event (#1093)
The `build` workflow needs to be simplified/split-up in preparation for https://github.com/winglang/wing/issues/256. The alert workflow doesn't need to run in forks, nor does it even need to run on push. This is one less thing to worry about in forks (check_run events should not be triggered by forks AFAIK)


*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.