feat: capture immutable values (#1076)
It is now possible to capture immutable primitives (strings, booleans,
numbers) and collections (array, set, map). This is done by serializing
them to JSON and including in the inflight bundle.

We expanded the bindings API such that it now accepts `resources` and
`data`. Resources are bound and `_toInflight`ed while data doesn't need
any bindings and it is JSON-serialized into the bundle.

### SDK support

In order to support this in the SDK, two things needed to happen. First,
the base `resource._bind()` method now only calls `_bind()` on the bound
resource (and adds a connection) if it has such as method. Otherwise, it
simply passes it through. This means that SDK users can technically pass
any object and the SDK will not be defensive about it. *We currently
assume "capture-safety" is the responsibility of the Wing compiler.*
Similarly, we only call `toInflight` if the object has this method.
Otherwise, we simply call `JSON.stringify`.

There's a special case for `Set` because JSON.stringify doesn't work out
of the box for sets.

The `Duration` type is serialized through its `toJSON` method into its
respective fields.

The SDK now takes a snapshot of the inflight code. To avoid ugly paths
in the code bundles, I am filtering out any lines that start with `//`.
Not ideal, but harmless (famous last words).

### Compiler support

* Update `capture.rs` to identify primitives and collections and
basically pass them as-is under `bindings`.
* Propagate a mutable `Diagnostics` vector into `capture.rs` so it is
possible to report errors. Specifically, if users try capture an
uncapturable type, an error is emitted.

### Testing

Added a couple of tests: `capture_collections.w` and
`capture_primitives.w` to cover the various types of immutable objects.



Fixes #1061



*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Hasan <45375125+hasanaburayyan@users.noreply.github.com>
Co-authored-by: yoav-steinberg <yoavs@monada.co>