chore: fork-based workflows (#1108)
Because the workflows won't run until merged to main. My fork at
https://github.com/MarkMcCulloh/wing has example of the workflow running
on a fork (AKA a place that doesn't have access to all the secrets). I
propose an admin force merges this PR.

### What's going on here:
- build workflow now uses pull_request_target in addition to push (for
`main` only)
- Ensured PROJEN_GITHUB_TOKEN is only used in places that would have
access to it (pushes to main on our repo)
- GITHUB_TOKEN should have all the permissions needed to do what we need
in forks
- Simplified the self-mutation logic and verified it works in a fork.
Eventually, the self mutator won't be specific to wingsdk :)

Fixes #256

*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.