fix(compiler): implicitly import types outside of namespace when importing jsii module (#1077)
Before this fix anything in the wingsdk needed to be contained within a single namespace. You couldn't, for example, have some class inherit from a class in another module. This is why we moved `FileBase` to be under `fs` instead of under `core`, because the JSII importer failed to import `fs.TextFile` because it included something from `core` in it.

This should resolve #834. 

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.