chore: move rfcs to docs (#1133)
Random idea. I don't see why the RFCs should be separate from the rest of the documentation, it'd be nice to be able to reference it from the beautiful rendered site.

I renamed the RFCs based on date so that they're ordered. Seems like we got away from doing "RFC Numbers" but I still wanted a way to order them somehow. Also separated the "stories" because they are kinda different from the others.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.