fix(compiler): using `print()` in bad places causes panics (#1122)
Fixes #1062

The implementation adds a new type `void` to our type system, used to represent the return type of functions that do not return a value. This type is not user-constructible (you cannot create a `void` value) and the `void` type is only equal to itself during type checking. In doing so we're able to remove a _lot_ of `unwrap` calls that corresponded to compiler panics. I've included a few examples in the `print_invalid` unit test.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.