rfc: wing sdk specification (#858)
[Rendered version](https://github.com/winglang/wing/blob/rybickic/sdk-rfc/docs/05-rfcs/2023-01-11-sdk-spec.md)

A draft of our first specification for the Wing SDK resources. While the SDK already has some resources and we can keep adding them on today, we would like to use this spec to make it consider how a larger set of SDK resources will look like, and how they will feel like to be used together, and what kinds of standard interfaces and patterns are needed.

For the first draft, our goal is to focus on API quality and consistency over API comprehensiveness. It's more important that our APIs are reliable, and that our APIs feel like they were all made with the same principles, and that they're all composable in intuitive ways -- than ensuring that every use case is supported.

Each resource has several parts:
- **Props** -  fields that can be passed to initialize and configure the resource. Props can be required or optional - but optional props must have a reasonable default which must be documented.
- **Preflight APIs** - properties and methods that can be accessed on the resource during preflight. Methods are often used to glue resources together, like `Bucket.on_upload(handler)`.
- **Inflight APIs** - properties and methods that can be accessed on the resource during inflight.

Related: #83

Closes #662

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.