fix: playground crashes due to memory leaks (#1199)
wingc is no longer a binary, now only a library. The cli and playground use it via WASM and pointer fun.
So, far this seems to increase the speed of `wing compile ...` by ~80%! It unfortunately slows down the playground a little bit, but now it no longer has a huge memory leak and won't just break after a while.

Right now the logic for `wingcInvoke` is duplicated for the cli and playground. `wingc` should eventually have a wrapper npm package that includes this logic. 

Added an "example" for wingc, which allows a similar workflow to the previous "cargo run". No doc updates needed, as it can be accessed via `nx dev` like before.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.