fix(playground): compiler errors missing (#1207)
regressions from https://github.com/winglang/wing/pull/1199

Writing to stderr from wingc was no longer causing an exception in the playground, which was hiding it. This does not affect the CLI, which uses the node WASI system rather than wasmer-js.

I also went ahead and made stderr more obvious, instead of showing it in where the user was expecting JS.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.