feat(sdk): experimental webapp resource
The `WebApp` resource can be used to create web applications that execute an inflight function in the browser.

    bring cloud;

    new cloud.WebApp(inflight () => {
      print("hello, wing!");
    });

If the above app is executed in the simulator, a localhost http endpoint will be started which will serve a simple website which executes the inflight function when it is loaded.

TODO:
- [ ] Add support for adding static assets to the site
- [ ] Support capturing resources (this requires some HTTP tunnel, so we are waiting for the `Api` resource).