feat(sdk): add timeout to sim function
Implement the timeout property for the simulator if functions inflight.
The timeout property is enforced by using the vm option to stop the
script execution after timeout.

On top of that a default timeout of 60s is set
if no timeout is set by the user.
This aligns with the rfc spec.