fix: unable to capture resources inside immutable data containers
This is an initial step towards merging `data` and `resources` in the binding system. We now allow any immutable data object to also include resources (e.g. an array of buckets, map of queues, etc).

To support this, we use duck-typing to determine if an object is a resource (e.g. it implements `_toInflight` and `_bind`), and call these methods in their relevant place respectively.

Fixes #1156