feat(sdk): extend function with timeout property  (#1220)
Closes: https://github.com/winglang/wing/issues/1204

This PR adds a timeout property to the `cloud.Function` class so that the execution time of a function can be limited. 
This property is defined in the rfc for the Function class. 
The RFC specifies a default value for timeout of 1 minute.

This PR adds a implementation for the timeout property for the simulator and aws function.

In the AWS function implementation the value of the timeout property is directly mapped to the cdktf property of the aws lambda function construct property.

The simulator implementation uses the node [vm options](https://nodejs.org/docs/latest-v16.x/api/vm.html#vmrunincontextcode-contextifiedobject-options) which offer a timeout property to cancel executions after a defined timeout. 

An alternative implementation of the timeout for the simulator has been not been evalutated because the implemented version resulted in the expected behaviour with a minimal effort. 
Side efects are not known yet. 
In case of a timeout in the simulator the User gets informed that a script execution has timed out.
This message could be improved to give the user feedback which inflight function has timed out.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.