fix(compiler): empty set initialization (#1262)
Empty Set (and MutSet) can now be initialized. 
We had a problem where tree-sitter would assume `{}` is a map (which is
our desired behavior) - this caused our parser to parse the right hand
expression of `let empty_set = Set<num>{};` as a map literal.
Added a special treatment for this case.

Also renamed one of our examples which had a typo in its name.

Fixes #1172 

*By submitting this pull request, I confirm that my contribution is made
under the terms of the
[Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

Co-authored-by: yoav-steinberg <yoavs@monada.co>