fix(docs): jsii-docgen tests install remote packages every time (#1268)
I found a few tests in jsii-docgen that would install some packages from the internet each time you run them. This makes the monorepo a bit tricky to work on if you have limited internet (on a plane, maybe). 😄  To fix this I updated those tests to just use one of the fixture libraries.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.