feat(examples): naive counting semaphore (#1260)
A naive implementation, not quite the right encapsulation I wanted.

For example, I feel as a distributed semaphore in a cloud environment should not be trapped in a while loop if fails to acquire a share (intentionally leave out the `acquire` implementation for now). Instead, it should be dispatched to something similar in a cloud resource for later pickup, like a queue, keep the async nature of being on cloud. However, I would like to get some early feedback to see if I'm getting the right vibe in my first winglang example.


*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.