feat(examples): naive distributed lock (#1248)
Another Wing example for inspiration!

See the source code for more explanation on limitations of this lock implementation. It's possible kind of lock might be practical if it was augmented with a timeout mechanism -- i.e. it internally tracked _who_ requested the lock and what time it was requested, and if the lock was not released after that amount of time, whoever requested the lock would be evicted/stopped and the lock would be released.

Another idea is to tweak this resource into a [semaphore](https://en.wikipedia.org/wiki/Semaphore_(programming)). A semaphore is basically a lock that can be acquired by up to N workers.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.