chore: strictly pin and update cdktf version in sdk (#1280)
The purpose of this is to stop snapshots from changing with new versions of cdktf. Originally I was just going to sanitize the snapshot, but I feel like it's entirely reasonable to pin cdktf very strictly. 

cdktf is pretty important and we may not want consumers of the SDK to freely choose the version.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.