chore: use pull_request instead of pull_request_target for builds (#1441)
We switched to pull_request_target a while back while making sure fork workflows would work. This made sense for the self-mutation workflow, but this isn't necessary for the build workflow. 
If triggered from a PR, the workflow does not use any secrets or attempt to write anything so it doesn't need those permissions. The only time secrets are read is on `push` to main.

The primary benefit of this change is that now we can easily test workflow changes in PRs. It's also just the generally recommended way for doing these workflows.


The explicit `ref` and `repository` settings were removed because pull_request_target didn't set it in the expected way. No longer needed, PRs will checkout the merge commit now.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.