feat(compiler): copy() and copy_mut() for builtin collections (#1427)
The main challenge here is having methods of a generic class that return another generic. I initially wanted to do a recursive style to have this work for any class but I didn't quite figure out a good way to handle it. I think when we add some sort of caching for this hydration process then we could add the recursion back in. 

For now I think it's okay to stick with our builtins, I think this is the most common situation.

*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.