feat: wasm language server and simplified build system (#1453)
<img src="https://user-images.githubusercontent.com/1237390/216994444-73b32bd8-6495-4994-9f08-cb4dc3922be5.png" width="600" />

The Wing cli now has a `lsp` command which starts the language server. With this change, there are no more native parts of our toolchain. There are several ramifications for this:
- Nx no longer has to worry about `native` vs `wasm`, everything is just `build`
- The github workflow similarly only has one build job now where it builds everything
- The existing wing-language-server project has been removed, most of the code migrated to wingc under the wls module
- vscode now just uses npx to resolve the wing cli for the language server, alternatively an env var.

In the process of the simplification, Fixed #1396 by removing the npm registry stuff from hangar.
Also Fixes #821 because I removed the rust tests from tree-sitter-wing so we actually run all the tests.

**With all these changes, the build time is cut in half!**

### Future Improvement
Testing the rust crates still requires native builds. To make it faster, it only does a debug build. I attempted to use a wasmtime to instead run the tests through that, but ran into too many hurdles so decided to leave it.

### Intentional regression
Removed yarn from hangar. Had issues forcing it to resolve the SDK tarball. This worked before because mocking the registry is pretty brute force. This is specifically a problem with hangar, not the usage of wing with yarn. In order for it to work, we have to hack around yarn to force certain resolutions. At that point, what's the value in testing with it?

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.